Terminando casos de uso. Ligero problema con el pattern del html

+++ src/main/java/com/example/demo/Controlador/Controlador.java
@@ -15,52 +15,86 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Controller
-public class Controlador{
+public class Controlador {
 
     @Autowired
     private DisciplinaServicio disciplinaServicio;
 
     /*
-    @RequestMapping("/")
-    public String disciplinaIH(){
-        return "disciplinaIH";
-    }
-    */
+     * @RequestMapping("/")
+     * public String disciplinaIH(){
+     * return "disciplinaIH";
+     * }
+     */
 
-    @GetMapping({"/disciplinaIH", "/"})
-    public String menuDisciplinas(Model modelo, @Param("busqueda") String busqueda){
+    @GetMapping({ "/disciplinaIH", "/" })
+    public String menuDisciplinas(Model modelo, @Param("busqueda") String busqueda) {
         modelo.addAttribute("disciplina", disciplinaServicio.mostrarDisciplinas(busqueda));
         modelo.addAttribute("busqueda", busqueda);
         return "disciplinaIH";
     }
 
     @PostMapping("/agregar")
-    public String procesa(HttpServletRequest request, Model modelo){
-        Disciplina disciplina = disciplinaServicio.agregaDisciplina(request.getParameter("disciplina"));
-        modelo.addAttribute("disciplina", disciplina);
-        return "redirect:/disciplinaIH";
+    public String procesa(HttpServletRequest request, Model modelo, RedirectAttributes redirectAttributes) {
+        Disciplina disciplina = new Disciplina();
+        disciplina.setNombre(request.getParameter("disciplina"));
+        if (disciplinaServicio.existeDisciplina(disciplina.getNombre())) {
+            modelo.addAttribute("error", true);
+            modelo.addAttribute("disciplina", disciplinaServicio.mostrarDisciplinas(""));
+            return "disciplinaIH";
+        } else {
+            String auxiliar = disciplina.getNombre();
+            auxiliar.trim();
+            auxiliar = auxiliar.substring(0, 1).toUpperCase() + auxiliar.substring(1).toLowerCase();
+            disciplinaServicio.agregaDisciplina(auxiliar);
+            modelo.addAttribute("disciplina", auxiliar);
+            redirectAttributes.addFlashAttribute("exitoAgrega", true);
+            return "redirect:/disciplinaIH";
+        }
     }
 
     @GetMapping("/disciplinaIH/editar/{id}")
-    public String formularioEditar(@PathVariable("id") Integer id, Model modelo){
+    public String formularioEditar(@PathVariable("id") Integer id, Model modelo) {
         modelo.addAttribute("disciplina", disciplinaServicio.consultarDisciplina(id));
         return "formulario";
     }
 
     @PostMapping("/disciplinaIH/{id}")
-    public String actualizarDisciplina(@PathVariable("id") Integer id, @ModelAttribute("disciplina") Disciplina disciplina, Model modelo){
+    public String actualizarDisciplina(@PathVariable("id") Integer id,
+            @ModelAttribute("disciplina") Disciplina disciplina, Model modelo, RedirectAttributes redirectAttributes) {
         Disciplina original = disciplinaServicio.consultarDisciplina(id);
-        original.setNombre(disciplina.getNombre());
-        disciplinaServicio.editarDisciplina(original);
-        return "redirect:/disciplinaIH";
+        String nombreOriginal = disciplina.getNombre();
+        if (disciplinaServicio.existeDisciplina(nombreOriginal)) {
+            modelo.addAttribute("error", true);
+            modelo.addAttribute("disciplina", disciplinaServicio.mostrarDisciplinas(""));
+            return "disciplinaIH";
+        } else {
+            String auxiliar = disciplina.getNombre();
+            auxiliar.trim();
+            auxiliar.toLowerCase();
+            auxiliar = auxiliar.substring(0, 1).toUpperCase() + auxiliar.substring(1);
+            disciplinaServicio.agregaDisciplina(auxiliar);
+            original.setNombre(auxiliar);
+            disciplinaServicio.editarDisciplina(original);
+            redirectAttributes.addFlashAttribute("exitoActualiza", true);
+            return "redirect:/disciplinaIH";
+        }
     }
 
     @GetMapping("/disciplinaIH/{id}")
-    public String eliminarDisciplina(@PathVariable Integer id){
-        disciplinaServicio.eliminarDisciplina(id);
-        return "redirect:/disciplinaIH";
+    public String eliminarDisciplina(@PathVariable Integer id, RedirectAttributes redirectAttributes, Model modelo) {
+        if (disciplinaServicio.existeDisciplina(disciplinaServicio.consultarDisciplina(id).getNombre())) {
+            disciplinaServicio.eliminarDisciplina(id);
+            redirectAttributes.addFlashAttribute("exitoElimina", true);
+            return "redirect:/disciplinaIH";
+        } else {
+            modelo.addAttribute("errorElimina", true);
+            return "disciplinaIH";
+            //No pasaría nunca por aquí por error en el mapping. :>
+        }
     }
-    
+
 }
